Add appengine oauth script from acd ouath service.
acd_cli had an outtage with their appengine service. Since the code is
shared via the appengine service, it made it difficult to try to spin
up a new instance of the service to work around the problem. This is
just here to ensure the code is available in the future. It should be
noted that Amazon has recently made whitelisting applications for
Amazon Drive invite only, which means unless you already have your app
up, it is unlikely they will approve your new instance of this
application.